server/requirements.txt
flask
python-dotenv
python-jose-cryptodome
flask-cors
six
Flask-SQLAlchemy==2.3.2